+++ node/clients/profile/profileV1.ts
@@ -5,7 +5,7 @@ import {
   RequestConfig,
 } from '@vtex/api'
 
-import { statusToError } from '../utils'
+import { statusToError } from '../../utils'
 
 const FIVE_SECONDS_MS = 5 * 1000
 
@@ -15,12 +15,12 @@ function mapAddressesObjToList(addressesObj: any): Address[] {
   )
 }
 
-export class ProfileClient extends JanusClient {
+export class ProfileClientV1 extends JanusClient {
   constructor(context: IOContext, options?: InstanceOptions) {
     super(context, {
       ...options,
       headers: {
-        ...(options && options.headers),
+        ...options?.headers,
         VtexIdClientAutCookie: context.authToken ?? '',
       },
       timeout: FIVE_SECONDS_MS,
@@ -36,9 +36,13 @@ export class ProfileClient extends JanusClient {
           extraFields: customFields,
         },
       }
-    )
+    ).then((profile) => {
+      profile.pii = false
 
-  public getUserAddresses = (user: CurrentProfile) =>
+      return profile
+    })
+
+  public getUserAddresses = (user: CurrentProfile, _: Profile) =>
     this.get(`${this.baseUrl}/${getUserIdentification(user)}/addresses`, {
       metric: 'profile-system-getUserAddresses',
     }).then(mapAddressesObjToList)
@@ -83,7 +87,8 @@ export class ProfileClient extends JanusClient {
 
   public updatePersonalPreferences = (
     user: CurrentProfile,
-    personalPreferences: PersonalPreferences
+    personalPreferences: PersonalPreferences,
+    _: Profile
   ) =>
     this.post(
       `${this.baseUrl}/${getUserIdentification(user)}/personalPreferences/`,