+++ node/directives/index.ts
@@ -3,11 +3,13 @@ import { WithSegment } from './withSegment'
 import { WithOrderFormId } from './withOrderFormId'
 import { ToVtexAssets } from './toVtexAssets'
 import { AuthorizationMetrics } from './authorizationMetrics'
+import { WithOwnerId } from './withOwnerId'
 
 export const schemaDirectives = {
   toVtexAssets: ToVtexAssets,
   withCurrentProfile: WithCurrentProfile,
   withSegment: WithSegment,
   withOrderFormId: WithOrderFormId,
+  withOwnerId: WithOwnerId,
   withAuthMetrics: AuthorizationMetrics,
 }