+++ node/resolvers/catalog/offer.ts
@@ -45,7 +45,7 @@ export const resolvers = {
         }
       )?.Value
 
-      return spotPrice || sellingPrice
+      return spotPrice ?? sellingPrice
     },
   },
 }